% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASpoly.data.R
\docType{class}
\name{GWASpoly.data-class}
\alias{GWASpoly.data-class}
\alias{GWASpoly.data}
\title{S4 class with genotype and phenotype data}
\description{
S4 class with genotype and phenotype data
}
\section{Slots}{

\describe{
\item{\code{map}}{data frame with marker,chrom,and position (either bp or cM)}

\item{\code{pheno}}{data frame of phenotypes}

\item{\code{geno}}{matrix (individuals x markers) of allele dosages (0,1,2,...ploidy)}

\item{\code{fixed}}{data frame of fixed effects}

\item{\code{ploidy}}{ploidy}
}}

