% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASpoly.R
\name{GWASpoly}
\alias{GWASpoly}
\title{Compute marker significance scores}
\usage{
GWASpoly(data, models, traits = NULL, params = NULL, n.core = 1, quiet = F)
}
\arguments{
\item{data}{Output from \code{set.K}}

\item{models}{Vector of model names}

\item{traits}{Vector trait names (by default, all traits)}

\item{params}{Optional list of params created by \code{set.params}}

\item{n.core}{Number of cores for parallel computing}

\item{quiet}{TRUE/FALSE whether to suppress output charting progress}
}
\value{
Variable of class \code{GWASpoly.fitted}
}
\description{
Compute marker significance scores
}
\details{
The following marker-effect models are available:
\itemize{
 \item "additive": Indicates the marker effect is proportional to the dosage of the alternate allele
 \item "X-dom": where X can be any integer between 1 and ploidy/2 and refers to the allele dosage needed for complete dominance (e.g., "1-dom" = simplex dominance, "2-dom" = duplex dominance).  The software tries both dominance patterns for a given dosage model, e.g., whether the reference or alternate allele is dominant
 \item "diplo-general": All heterozygotes have the same effect
 \item "diplo-additive": All heterozygotes have the same effect, constrained to be halfway between the homozygous effects
 \item "general": There are no constraints on the effects of the different dosage levels
 }
To specify additional model parameters, such as the inclusion of fixed effects (Q matrix) and the minimum minor allele frequency, use \code{set.params}
}
