% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASpoly.thresh.R
\docType{class}
\name{GWASpoly.thresh-class}
\alias{GWASpoly.thresh-class}
\alias{GWASpoly.thresh}
\title{S4 class with results from genome-wide scan and detection threshold}
\description{
S4 class with results from genome-wide scan and detection threshold
}
\section{Slots}{

\describe{
\item{\code{map}}{data frame with marker,chrom,and position (either bp or cM)}

\item{\code{pheno}}{data frame of phenotypes}

\item{\code{geno}}{matrix with allele dosages}

\item{\code{fixed}}{data frame of fixed effects}

\item{\code{ploidy}}{ploidy}

\item{\code{K}}{covariance matrix for polygenic effect}

\item{\code{scores}}{-log10(p) results}

\item{\code{effects}}{estimated marker effects}

\item{\code{params}}{parameters used for the analysis}

\item{\code{threshold}}{thresholds for significance}
}}

