% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keff.R
\name{Keff}
\alias{Keff}
\title{Effective number of markers}
\usage{
Keff(r2, alpha)
}
\arguments{
\item{r2}{matrix of squared correlations}

\item{alpha}{significance level}
}
\value{
numeric
}
\description{
Effective number of markers
}
\references{
Moskvina V, Schmidt KM (2008) On multiple-testing correction in genome-wide association studies. Genetic Epidemiology 32:567-573. doi:10.1002/gepi.20331
}
\keyword{internal}
