% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.plot.R
\name{LD.plot}
\alias{LD.plot}
\title{Plot LD vs distance}
\usage{
LD.plot(data, max.pair = 10000, dof = 8, max.loci = NULL, position = "bp")
}
\arguments{
\item{data}{variable inheriting from class \code{\link{GWASpoly}}}

\item{max.pair}{maximum number of r2 pairs for the spline}

\item{dof}{degrees of freedom for the spline}

\item{max.loci}{maximum number of markers to use per chromosome}

\item{position}{"bp" or "cM"}
}
\value{
ggplot2 object
}
\description{
Plot LD vs distance
}
\details{
A monotone decreasing, convex spline is fit using R package \code{scam}.
}
