% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.plot.R
\name{manhattan.plot}
\alias{manhattan.plot}
\title{Create Manhattan plot}
\usage{
manhattan.plot(data, traits = NULL, models = NULL, chrom = NULL)
}
\arguments{
\item{data}{Variable of class \code{GWASpoly.fitted}}

\item{traits}{Vector of trait names (by default, all traits plotted)}

\item{models}{Vector of model names (by default, all models plotted)}

\item{chrom}{optional, to plot only one chromosome}
}
\value{
ggplot2 object
}
\description{
Create Manhattan plot
}
\details{
Results for the ref and alt versions of the dominance model are combined. If \code{data} is the output from \code{\link{set.threshold}}, then the threshold is displayed as a horizontal dashed line when \code{models} contains a single model. Because the threshold varies between models, it is not drawn when multiple models are included. Although the ref and alt versions of each dominance model are slightly different (as seen with \code{\link{qq.plot}}), they are treated as a single model for the Manhattan plot, and the average threshold is shown.
}
