% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF2dosage.R
\name{VCF2dosage}
\alias{VCF2dosage}
\title{Prepare GWASpoly dosage input file from VCF file}
\usage{
VCF2dosage(
  VCF.file,
  dosage.file,
  geno.code,
  ploidy,
  samples = NULL,
  min.DP = 1,
  max.missing,
  min.minor = 5
)
}
\arguments{
\item{VCF.file}{VCF filename to read in}

\item{dosage.file}{name of CSV file to output with allele dosage}

\item{geno.code}{genotype code in the FORMAT field: either "GT" or "DS" (see Details)}

\item{ploidy}{ploidy}

\item{samples}{optional vector of sample names, to export subset of the population}

\item{min.DP}{minimum per sample depth (DP) to export genotype. Default is 1, for no filtering.}

\item{max.missing}{threshold for missing data per marker, as a proportion.}

\item{min.minor}{minimum number of samples with the minor allele. Default is 5.}
}
\description{
Prepare GWASpoly dosage input file from VCF file
}
\details{
The "GT" option for \code{geno.code} is the posterior maximum genotype (e.g., 0/0/1/1), while "DS" represents the posterior mean dosage of the alternate allele. (At present, only bi-allelic variants are supported.) The VCF file must conform to version 4.1 of the VCF specification or later.
}
