% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.GWASpoly.R
\name{write.GWASpoly}
\alias{write.GWASpoly}
\title{Write results to file}
\usage{
write.GWASpoly(data, trait, filename, what = "scores", delim = ",")
}
\arguments{
\item{data}{Variable of class \code{GWASpoly.fitted}}

\item{trait}{Trait name}

\item{filename}{Filename}

\item{what}{Either "scores" or "effects"}

\item{delim}{Delimiter to use in the output file (default is comma)}
}
\description{
Write results to file
}
\details{
Score = -log10(p). Effect = marker effect (not available for the general and diplo-general models).
}
