% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.QTL.R
\name{fit.QTL}
\alias{fit.QTL}
\title{Test markers as QTL under backward elimination}
\usage{
fit.QTL(data, trait, qtl, fixed = NULL)
}
\arguments{
\item{data}{variable inheriting from class \code{\link{GWASpoly.K}}}

\item{trait}{name of trait}

\item{qtl}{data frame to specify the multi-QTL model (see Details)}

\item{fixed}{data frame to specify the fixed effects (see Details)}
}
\value{
data frame with partial r2 and p-values
}
\description{
Test markers as QTL under backward elimination
}
\details{
\code{qtl} is a data frame with columns "Marker" and "Model", where each row corresponds to a QTL. \code{fixed} is a data frame with columns "Effect" and "Type": the first column is the name of the effect, which must match a column in the phenotype input file, and the second column is either "factor" or "numeric". The p-value and R2 for each marker are based on the likelihood ratio test under backward elimination, comparing the deviance to the chi-squared distribution.
}
