% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.QTL.R
\name{get.QTL}
\alias{get.QTL}
\title{Extract significant QTL}
\usage{
get.QTL(data, traits = NULL, models = NULL)
}
\arguments{
\item{data}{Output from \code{set.threshold}}

\item{traits}{Vector of trait names (by default, all traits)}

\item{models}{Vector of model names (by default, all models)}
}
\value{
Data frame with results for significant markers
}
\description{
Output a table with all significant markers
}
\details{
Score = -log10(p). Effect = marker effect (not available for the general and diplo-general models).
}
