% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASpoly.fitted.R
\docType{class}
\name{GWASpoly.fitted-class}
\alias{GWASpoly.fitted-class}
\alias{GWASpoly.fitted}
\title{S4 class with results from genome-wide scan}
\description{
S4 class with results from genome-wide scan
}
\section{Slots}{

\describe{
\item{\code{map}}{data frame with marker,chrom,and position (either bp or cM)}

\item{\code{pheno}}{data frame of phenotypes}

\item{\code{geno}}{matrix with allele dosages}

\item{\code{fixed}}{data frame of fixed effects}

\item{\code{ploidy}}{ploidy}

\item{\code{K}}{list of covariance matrices}

\item{\code{scores}}{-log10(p) results}

\item{\code{effects}}{estimated marker effects}

\item{\code{params}}{parameters used for the analysis}
}}

