% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASpoly.K.R
\docType{class}
\name{GWASpoly.K-class}
\alias{GWASpoly.K-class}
\alias{GWASpoly.K}
\title{S4 class with genotypes, phenotypes, and polygenic covariance}
\description{
S4 class with genotypes, phenotypes, and polygenic covariance
}
\section{Slots}{

\describe{
\item{\code{map}}{data frame with columns Marker,Chrom,Position,Ref,Alt}

\item{\code{pheno}}{data frame of phenotypes}

\item{\code{geno}}{matrix with allele dosages}

\item{\code{fixed}}{data frame of fixed effects}

\item{\code{ploidy}}{ploidy}

\item{\code{K}}{list of covariance matrices (one for each chromosome)}
}}

