% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.plot.R
\name{manhattan.plot}
\alias{manhattan.plot}
\title{Create Manhattan plot}
\usage{
manhattan.plot(data, traits = NULL, models = NULL, y.max = NULL)
}
\arguments{
\item{data}{Variable of class \code{GWASpoly.fitted}}

\item{traits}{Vector of trait names (by default, all traits plotted)}

\item{models}{Vector of model names (by default, all models plotted)}
}
\value{
ggplot2 object
}
\description{
Create Manhattan plot
}
\details{
Results for the ref and alt versions of the dominance model are combined
}
