% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.GWASpoly.R
\name{read.GWASpoly}
\alias{read.GWASpoly}
\title{Read in marker and phenotype data}
\usage{
read.GWASpoly(ploidy, pheno.file, geno.file, format, n.traits, delim = ",")
}
\arguments{
\item{ploidy}{Ploidy (e.g., 2 for diploid, 4 for tetraploid)}

\item{pheno.file}{Name of the phenotype file}

\item{geno.file}{Name of the genotype file}

\item{format}{Format for the marker data. See details.}

\item{n.traits}{Number of traits}

\item{delim}{Character to indicate the delimiter in the data files (e.g., "," for csv, "\\t" for tab-delimited)}
}
\value{
Variable of class \code{GWASpoly.data}
}
\description{
Read in marker and phenotype data
}
\details{
The first column of the phenotype file contains the genotype identifier, columns 2 through (n.traits + 1) contain trait values, and subsequent columns contain the levels (for factors) or numeric values (for covariates) of any fixed effects.  
The first three columns of the genotype file are (1) marker name, (2) chromosome, and (3) position.  Subsequent columns contain the marker data for each individual in the population.  
Marker data can be coded in one of three formats: 
\itemize{
\item "numeric": markers are coded based on the dosage of the alternate allele, taking on integer values between 0 and ploidy (fractional values not allowed)
\item "AB": e.g., AAAB, ABBB for tetraploids
\item "ACGT": e.g., AAAT, GGCC for tetraploids
}
Only bi-allelic markers are allowed.  Missing marker data will be imputed with the population mode (most frequent value) for each marker.
}
