% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.K.R
\name{set.K}
\alias{set.K}
\title{Set covariance matrix for polygenic effect}
\usage{
set.K(data, K = NULL)
}
\arguments{
\item{data}{Output from \code{read.GWASpoly}}

\item{K}{Optional: user-supplied matrix}
}
\value{
Variable of class \code{GWASpoly.K}
}
\description{
Set covariance matrix for polygenic effect
}
\details{
By default, K is computed as $K=MM'$, where M is the centered genotype matrix (lines x markers).  For GWAS, the overall scaling of K is irrelevant.  At present, K is scaled such that the mean of its diagonal elements is 1.  Alternatively, the user can supply any positive semidefinite K (with row.names that match the genotype identifiers).
}
