% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.K.R
\name{set.K}
\alias{set.K}
\title{Set covariance matrix for polygenic effect}
\usage{
set.K(data, K = NULL, n.core = 1, LOCO = TRUE)
}
\arguments{
\item{data}{Output from \code{read.GWASpoly}}

\item{K}{Optional: user-supplied matrix}

\item{n.core}{Number of cores for parallel computing}

\item{LOCO}{TRUE/FALSE, whether to use leave-one-chromosome-out}
}
\value{
Variable of class \code{GWASpoly.K}
}
\description{
Set covariance matrix for polygenic effect
}
\details{
When \code{LOCO} = TRUE, K is computed for each chromosome as $K=MM'$, where M is the centered genotype matrix (lines x markers), and scaled to have unit diagonal (the overall scaling is not important for GWAS). When \code{LOCO} = FALSE, a single K matrix is computed for all markers (this is not recommended but provided for legacy reasons). Alternatively, the user can supply their own positive semidefinite K, with row.names that match the genotype identifiers (this option cannot be used with LOCO).
}
