% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.threshold.R
\name{set.threshold}
\alias{set.threshold}
\title{Set the significance threshold}
\usage{
set.threshold(data, method, level = 0.05, n.permute = 1000, n.core = 1)
}
\arguments{
\item{data}{Variable of class \code{GWASpoly.fitted}}

\item{method}{One of the following: "Bonferroni","FDR","permute"}

\item{level}{Genome-wide false positive rate for the Bonferroni or permutation methods; false discovery rate for method FDR}

\item{n.permute}{Number of permutations for method "permute"}

\item{n.core}{Number of cores to use for multicore processing}
}
\value{
Variable of class \code{GWASpoly.thresh}
}
\description{
Set the significance threshold
}
\details{
The FDR method is based on version 1.30.0 of the qvalue package
}
