% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF2dosage.R
\name{VCF2dosage}
\alias{VCF2dosage}
\title{Convert VCF to dosage file}
\usage{
VCF2dosage(
  VCF.file,
  dosage.file,
  geno.code,
  ploidy,
  samples = NULL,
  min.DP = 1,
  max.missing,
  min.minor = 5
)
}
\arguments{
\item{VCF.file}{VCF filename (can be gzipped)}

\item{dosage.file}{CSV filename to output with allele dosage}

\item{geno.code}{genotype code in the FORMAT field: "GT" of "DS"}

\item{ploidy}{ploidy}

\item{samples}{optional vector of sample names, to export subset of the population}

\item{min.DP}{minimum per sample depth (DP) to export genotype. Default is 1, for no filtering.}

\item{max.missing}{threshold for missing data per marker, as a proportion.}

\item{min.minor}{minimum number of samples with the minor allele. Default is 5.}
}
\description{
Convert VCF to dosage file
}
\details{
Only bi-allelic variants supported. The "GT" option for \code{geno.code} is the posterior maximum genotype (e.g., 0/0/1/1). "DS" represents the posterior mean dosage of the alternate allele. VCF file must conform to 4.1 or later.
}
